LPS-74544 Only apply existing code to *.jsp
